Move generated files back from target folder, fix .gitignore
