+++ cocas/main.py
@@ -44,7 +44,7 @@ def main():
     isa = importlib.import_module(f'targets.{isa_name}', 'cocas')
 
     library_macros = read_mlb(
-        str(pathlib.Path(__file__).parent.joinpath(f'targets/{isa_name}/standard.mlb').absolute()), isa)
+        str(pathlib.Path(__file__).parent.joinpath(f'standard.mlb').absolute()))
     objects = []
 
     for filepath in args.sources:
@@ -71,9 +71,9 @@ def main():
             # Hash of source file name
             # Remove comments
             macro_expanded_input_stream = process_macros(input_stream, library_macros,
-                                                         str(pathlib.Path(filepath).absolute()), isa)
+                                                         str(pathlib.Path(filepath).absolute()))
             # print(macro_expanded_input_stream)
-            r = build_ast(macro_expanded_input_stream, str(pathlib.Path(filepath).absolute()), isa)
+            r = build_ast(macro_expanded_input_stream, str(pathlib.Path(filepath).absolute()))
             obj = assemble(r, isa)
 
             objects.append(obj)
@@ -82,7 +82,7 @@ def main():
             return 1
 
     try:
-        data, code_locations = link(objects, isa)
+        data, code_locations = link(objects)
     except CdmLinkException as e:
         log_error(str(CdmExceptionTag.LINK), e.message)
         return 1
