--- cocas/targets/cdm8e/__init__.py
@@ -1,6 +1,0 @@
-import cocas.targets.cdm8e.AsmLexer as AsmLexer
-import cocas.targets.cdm8e.AsmParser as AsmParser
-import cocas.targets.cdm8e.AsmParserVisitor as AsmParserVisitor
-import cocas.targets.cdm8e.MacroLexer as MacroLexer
-import cocas.targets.cdm8e.MacroParser as MacroParser
-import cocas.targets.cdm8e.MacroVisitor as MacroVisitor
